Move skeleton-query-wizard.ts to /local-queries
